Fix #2730 Fix #2731 Fix #2732 Fix #2733: ui.md counts, main.rs split, self-expiring dead-code expects, shared Ruffle device
TD3-2026-08-12-03 (#2730) — `docs/engine/ui.md` said the executable adds
three winit-translation tests; `ui_input.rs` has four. The crate half the
issue certified as "exactly right (verified: 16 passed, 3 ignored)" was
also wrong: the suite runs 24 by measurement, 26 after this commit's two
additions. Both numbers corrected against a live run rather than the
issue text.

TD1-2026-08-12-01 (#2731) — `main.rs` was 2,147 LOC by the time this
landed (2,054 when filed). Split along the axis the issue prescribes:
`App::render_one_frame` → `app_frame.rs`, `impl ApplicationHandler for
App` → `app_events.rs`, joining the existing `app_step.rs`. `App`, its
construction and the debug-UI bridge stay in `main.rs`. 2,147 → 679 LOC,
with the new modules at 545 and 988; nothing is left over the threshold.

The move is verbatim — diffing the extracted ranges against the new files
shows one deliberate change, `render_one_frame` widening to `pub(crate)`
for the cross-module call.

What the split silently broke, and how: three source-position tests
`include_str!("main.rs")` and search for code that moved. They did not
fail. `find` located each needle's own literal inside the test module and
every assertion passed while pinning nothing — the same self-match trap
that shows up whenever a test greps the file it lives in. Both test
modules moved to the files they now describe, and `boot.rs`'s
`overlay_open_arms_the_tracker_lazily` was repointed at `app_events.rs`.
Verified live afterwards by perturbing the pinned call and watching the
assertion fail.

TD8-2026-08-12-04 (#2732) — filed as tracking-only ("no action now"), but
the check it asks for can be made permanent instead of deferred to the
next sweep. The compiler says exactly `Attack`, `Block` and `Pause` lack
producers — `Inventory` is bound to Tab, so the finding's list of four was
already one stale. Replaced the blanket enum-level `#[allow(dead_code)]`
with per-variant `#[expect(dead_code, reason = …)]`. Two gains: a blanket
allow silently absorbs the *next* unproduced variant, and `expect` is
self-expiring — the build warns "unfulfilled lint expectation" the moment
a variant gains a producer, so whoever wires up mouse/gamepad deletes the
attribute instead of inheriting it. Confirmed by temporarily binding
`Attack` and watching the warning fire.

CONC-D7-UI-05 (#2733) — every `SwfPlayer` built its own wgpu instance,
adapter, device and queue under `block_on` on the winit main-loop thread,
so each menu load created a second live Vulkan device, and a menu swap
transiently held two Ruffle devices plus the engine's. `Descriptors` is
now a process-wide singleton, which is the sharing model Ruffle itself
assumes (`WgpuRenderBackend::new` takes an `Arc<Descriptors>` so several
players can share one device, each with its own `TextureTarget`).

The trade is stated in the code: the device now outlives
`UiManager::close()` and lives for the process. That is the point — a menu
can reopen at any time — but it does retain one idle logical device after
the last menu closes. A failed creation is deliberately not cached, so a
transient failure can't permanently disable the UI.

The caching rule is extracted as a generic `get_or_try_init` so the
default suite can pin it on a machine with no Vulkan device: init runs
once and is shared, and a failure leaves the slot empty for the retry.

Pre-existing unrelated failure: `fire_lights::derived_reach_is_currently
_oversized_pending_a_unit_correct_derivation` (canary, fails on main).
