+++ byroredux/src/main.rs
@@ -10,6 +10,8 @@
 static DHAT_ALLOC: dhat::Alloc = dhat::Alloc;
 
 mod anim_convert;
+mod app_events;
+mod app_frame;
 mod app_step;
 mod asset_provider;
 mod bench;
@@ -48,29 +50,16 @@ mod ui_input;
 
 use anyhow::Result;
 use byroredux_core::console::CommandRegistry;
-use byroredux_core::ecs::{
-    ActiveCamera, Camera, DebugStats, DeltaTime, EngineConfig, Scheduler, ScratchTelemetry,
-    SkinCoverageStats, TotalTime, World,
-};
+use byroredux_core::ecs::{Scheduler, World};
 use byroredux_core::settings::SettingsRegistry;
 use byroredux_core::string::StringPool;
-use byroredux_platform::window::{self, WindowConfig};
-use byroredux_renderer::vulkan::context::{DrawCommand, FrameInputs};
-use byroredux_renderer::vulkan::GpuUploadCtx;
-use byroredux_renderer::{ImageSpaceModifierView, RendererConfig, VulkanContext};
-use byroredux_ui::{ScaleformHostDispatch, UiManager};
+use byroredux_renderer::vulkan::context::DrawCommand;
+use byroredux_renderer::{RendererConfig, VulkanContext};
+use byroredux_ui::UiManager;
 use std::collections::HashMap;
 use std::time::Instant;
-use winit::application::ApplicationHandler;
-use winit::event::{DeviceEvent, DeviceId, ElementState, WindowEvent};
-use winit::event_loop::ActiveEventLoop;
-use winit::keyboard::{KeyCode, PhysicalKey};
-use winit::window::{CursorGrabMode, Window, WindowId};
-
-use crate::components::InputState;
-use crate::helpers::world_resource_set;
-use crate::render::build_render_data;
-use crate::systems::{compute_underwater_params, toggle_player_mode};
+use winit::event::WindowEvent;
+use winit::window::{CursorGrabMode, Window};
 
 fn main() -> Result<()> {
     boot::run()
@@ -539,1390 +528,6 @@ impl App {
 // can be called from both the `WindowEvent::RedrawRequested` arm (now
 // a no-op) and the `about_to_wait` tick. Trait impl reopens after the
 // inherent block below.
-impl App {
-    /// Phase 14 — pulled out of the original `WindowEvent::RedrawRequested`
-    /// arm so the game loop can call it directly from `about_to_wait`
-    /// instead of routing through `request_redraw()` → wait for the
-    /// compositor's frame-callback → `RedrawRequested`. On Wayland +
-    /// winit 0.30, that round-trip gates the engine at the
-    /// compositor's pace (~18 FPS in the observed Sleeping Giant Inn
-    /// reading) regardless of `ControlFlow::Poll`. Drawing from
-    /// `about_to_wait` bypasses the gate; combined with MAILBOX
-    /// present mode the compositor still vsyncs presentation but
-    /// the render loop runs uncapped.
-    fn render_one_frame(&mut self, event_loop: &ActiveEventLoop) {
-        // Phase 15 — bracket render_one_frame in three phases
-        // so the egui Metrics panel can pin which one of
-        // pre-draw / draw_frame call / post-draw is hiding the
-        // ~30 ms we still can't see (Phase-14 surfaced
-        // render_one_frame's total wall as ~47 ms while the
-        // GPU + per-call CPU brackets sum to ~18 ms).
-        let rof_pre_t0 = Instant::now();
-        let mut rof_pre_draw_ns: u64 = 0;
-        let mut rof_draw_call_ns: u64 = 0;
-        // Phase 4 — populate the panel snapshot from the
-        // World, run egui (gets `PanelOutputs` back), apply
-        // those outputs, then stash the FullOutput +
-        // egui::Context for the renderer to consume.
-        //
-        // #1376: build_debug_ui_snapshot deep-clones two BTreeMaps +
-        // a Vec of Strings every frame. Gate those diagnostics on
-        // `visible`; the interaction prompt is the only snapshot field
-        // populated while the operator overlay is hidden.
-        let snapshot = if self.debug_ui.as_ref().is_some_and(|ui| ui.visible) {
-            build_debug_ui_snapshot(&self.world, self.debug_ui_refresh_entities)
-        } else {
-            byroredux_debug_ui::PanelSnapshot {
-                interaction_prompt: build_interaction_prompt(&self.world),
-                ..Default::default()
-            }
-        };
-        self.debug_ui_refresh_entities = false;
-
-        let (egui_frame, outputs) =
-            if let (Some(ref mut ui), Some(win)) = (self.debug_ui.as_mut(), self.window.as_ref()) {
-                let outputs = ui.run(win, &snapshot);
-                let frame = ui.take_output().map(|out| (ui.egui_ctx.clone(), out));
-                (frame, outputs)
-            } else {
-                (None, byroredux_debug_ui::PanelOutputs::default())
-            };
-
-        apply_debug_ui_outputs(
-            &mut self.world,
-            outputs,
-            &mut self.debug_ui_refresh_entities,
-            self.debug_ui.as_mut(),
-        );
-
-        if let Some(ref mut ctx) = self.renderer {
-            if let Some((egui_ctx, output)) = egui_frame {
-                ctx.submit_egui_frame(egui_ctx, output);
-            }
-            let is_benching = self.bench_frames_target.is_some();
-
-            let brd_t0 = Instant::now();
-            let frame = build_render_data(
-                &self.world,
-                &mut self.draw_commands,
-                &mut self.water_commands,
-                &mut self.gpu_lights,
-                &mut self.gpu_fog_volumes,
-                &mut self.light_sort_scratch,
-                &mut self.bone_world,
-                &mut self.skin_offsets,
-                &mut self.skin_slot_pool,
-                &mut self.material_table,
-                ctx.particle_quad_handle,
-            );
-            if is_benching {
-                self.bench_build_render_ns += brd_t0.elapsed().as_nanos() as u64;
-            }
-
-            {
-                let mut tlm = self.world.resource_mut::<ScratchTelemetry>();
-                tlm.materials_unique = self.material_table.unique_user_count();
-                tlm.materials_interned = self.material_table.interned_count();
-                tlm.materials_overflow = self.material_table.overflow_count();
-            }
-            // #1428 — catch any frame where we silently degraded over-cap
-            // materials to slot 0 before the Once-gated warn fires again.
-            // Only fires in debug; the `ctx.scratch` console command surfaces the
-            // per-frame count in all builds.
-            debug_assert_eq!(
-                self.material_table.overflow_count(),
-                0,
-                "MaterialTable overflow: {} intern call(s) fell back to the \
-                 neutral-default slot 0 (MAX_MATERIALS={cap}). Run \
-                 `ctx.scratch` to confirm; consider raising MAX_MATERIALS in \
-                 scene_buffer/constants.rs if this cell genuinely needs it.",
-                self.material_table.overflow_count(),
-                cap = byroredux_renderer::MAX_MATERIALS,
-            );
-
-            let defer_geometry_rebuild = self
-                .streaming
-                .as_ref()
-                .is_some_and(streaming::WorldStreamingState::geometry_batch_in_progress);
-            if ctx.mesh_registry.is_geometry_dirty() && !defer_geometry_rebuild {
-                if let Err(e) = ctx.mesh_registry.rebuild_geometry_ssbo(
-                    &ctx.device,
-                    ctx.allocator.as_ref().unwrap(),
-                    &ctx.graphics_queue,
-                    ctx.transfer_pool,
-                    ctx.device_caps.ray_query_supported,
-                ) {
-                    log::warn!("Failed to rebuild geometry SSBO: {e}");
-                }
-            }
-            if ctx.mesh_registry.is_geometry_dirty() {
-                // Cell/LOD application appends CPU-side global geometry a few
-                // meshes at a time. Until the coherent batched rebuild lands,
-                // keep those tail ranges out of raster and TLAS; otherwise the
-                // old bound SSBO is indexed past its end. Existing resident
-                // meshes keep rendering while the transaction progresses.
-                for command in &mut self.draw_commands {
-                    if !ctx.mesh_registry.is_geometry_resident(command.mesh_handle) {
-                        command.in_raster = false;
-                        command.in_tlas = false;
-                    }
-                }
-                self.water_commands
-                    .retain(|command| ctx.mesh_registry.is_geometry_resident(command.mesh_handle));
-            }
-
-            ctx.build_global_blas_for_draws(&self.draw_commands);
-
-            world_resource_set::<DebugStats>(&self.world, |s| {
-                s.draw_command_count = self.draw_commands.len() as u32;
-            });
-
-            // Tick and render the UI overlay (Ruffle SWF player).
-            let ui_t0 = Instant::now();
-            let mut ui_tex = None;
-            if let Some(ref mut ui) = self.ui_manager {
-                let dt = self
-                    .world
-                    .try_resource::<DeltaTime>()
-                    .map(|d| d.0 as f64)
-                    .unwrap_or(1.0 / 60.0);
-                let ui_w = ui.width;
-                let ui_h = ui.height;
-                ui.tick(dt);
-
-                // Consume what the menu asked of the host (#2714). The bridge
-                // is drain-based by design and had no consumer outside its own
-                // tests, so every ActionScript call was retained for the life
-                // of the menu. Draining here is what keeps the queue at its
-                // natural depth; `MAX_QUEUED_CALLS` is only the backstop for
-                // when this does not run.
-                //
-                // Acting on the calls is M48 work — routing them into quest /
-                // inventory / player state needs those systems' menu contracts
-                // to exist first. What the engine can honestly do today is
-                // consume them and say which ones it cannot answer, which
-                // turns the bridge's `unknown_methods()` set from a test-only
-                // observation into a live one.
-                for call in ui.drain_host_calls() {
-                    log::debug!(
-                        "Scaleform host call #{} {} -> {} ({:?}, {} arg(s))",
-                        call.sequence,
-                        call.transport_method,
-                        call.method,
-                        call.dispatch,
-                        call.arguments.len(),
-                    );
-                    if matches!(
-                        call.dispatch,
-                        ScaleformHostDispatch::Unknown | ScaleformHostDispatch::MissingResponse
-                    ) && self.ui_reported_host_methods.insert(call.method.clone())
-                    {
-                        log::warn!(
-                            "Scaleform menu '{}' called host method '{}' ({:?}) — \
-                             no engine handler is registered, so the menu received Null",
-                            ui.menu_name,
-                            call.method,
-                            call.dispatch,
-                        );
-                    }
-                }
-
-                if let Some(pixels) = ui.render() {
-                    if let Some(handle) = self.ui_texture_handle {
-                        let allocator = ctx.allocator.as_ref().unwrap();
-                        let upload_ctx = GpuUploadCtx {
-                            device: &ctx.device,
-                            allocator,
-                            queue: &ctx.graphics_queue,
-                            command_pool: ctx.transfer_pool,
-                        };
-                        if let Err(e) = ctx
-                            .texture_registry
-                            .update_rgba(upload_ctx, handle, ui_w, ui_h, pixels)
-                        {
-                            log::error!("UI texture update failed: {e:#}");
-                        }
-                        ui_tex = Some(handle);
-                    }
-                } else if self.ui_texture_handle.is_some() {
-                    ui_tex = self.ui_texture_handle;
-                }
-            }
-            if is_benching {
-                self.bench_ui_ns += ui_t0.elapsed().as_nanos() as u64;
-            }
-
-            let is_interior = self
-                .world
-                .try_resource::<crate::components::CellLightingRes>()
-                .is_some_and(|l| l.is_interior);
-            let clear_color = if is_interior {
-                [0.0, 0.0, 0.0, 1.0]
-            } else {
-                byroredux_core::types::Color::CORNFLOWER_BLUE.as_array()
-            };
-            let render_t0 = Instant::now();
-            let mut frame_timings = Some(byroredux_renderer::FrameTimings::default());
-            let pending = self.skin_slot_pool.drain_pending(
-                byroredux_renderer::vulkan::scene_buffer::MAX_PENDING_BIND_INVERSE_UPLOADS_PER_FRAME,
-            );
-            // #1791 / D6-01 — mirror of `pending_with_data`'s (slot, entity)
-            // pairs, kept alive so a `draw_frame` early return (see the
-            // `skin_dispatch_ran` check below) can requeue exactly what was
-            // about to be uploaded. Deliberately NOT the raw `pending` drain:
-            // an entry filtered out here (its `SkinnedMesh` is already gone)
-            // must stay dropped, not come back through the requeue path.
-            let mut pending_for_requeue: Vec<(u32, byroredux_core::ecs::EntityId)> =
-                Vec::with_capacity(pending.len());
-            let pending_with_data: Vec<(u32, Vec<[[f32; 4]; 4]>)> = pending
-                .into_iter()
-                .filter_map(|(slot, entity)| {
-                    self.world
-                        .get::<byroredux_core::ecs::SkinnedMesh>(entity)
-                        .map(|skin| {
-                            let mut padded: Vec<[[f32; 4]; 4]> = skin
-                                .bind_inverses
-                                .iter()
-                                .map(|m| m.to_cols_array_2d())
-                                .collect();
-                            padded.resize(
-                                byroredux_core::ecs::components::MAX_BONES_PER_MESH,
-                                [
-                                    [1.0, 0.0, 0.0, 0.0],
-                                    [0.0, 1.0, 0.0, 0.0],
-                                    [0.0, 0.0, 1.0, 0.0],
-                                    [0.0, 0.0, 0.0, 1.0],
-                                ],
-                            );
-                            pending_for_requeue.push((slot, entity));
-                            (slot, padded)
-                        })
-                })
-                .collect();
-            // Phase 15 — close pre-draw bracket, open draw-call.
-            rof_pre_draw_ns = rof_pre_t0.elapsed().as_nanos() as u64;
-            let rof_draw_call_t0 = Instant::now();
-            let dof = byroredux_renderer::DofView {
-                aperture: frame.aperture,
-                focus_dist: frame.focus_dist,
-                cam_right: frame.cam_right,
-                cam_up: frame.cam_up,
-                cam_forward: frame.cam_forward,
-                proj_mat: frame.proj_mat,
-                camera_near: frame.camera_near,
-                camera_far: frame.camera_far,
-                camera_fov_y: frame.camera_fov_y,
-            };
-            // REND-#1451 — push live point/spot attenuation tuning
-            // (LightTuning resource, mutated by the `light.atten`
-            // console command) into the renderer so the controlled
-            // bench can sweep the knee / A/B the legacy model without a
-            // rebuild. Absent resource → renderer keeps its defaults.
-            if let Some(lt) = self.world.try_resource::<crate::components::LightTuning>() {
-                ctx.light_atten_knee = lt.knee_frac;
-                ctx.light_atten_legacy = lt.legacy;
-            }
-            let frame_time_delta_ms = self
-                .world
-                .try_resource::<DeltaTime>()
-                .map_or(1000.0 / 60.0, |delta| delta.0 * 1000.0);
-            let image_space_modifier = self
-                .world
-                .try_resource::<byroredux_scripting::CinematicPresentationState>()
-                .map_or_else(ImageSpaceModifierView::default, |state| {
-                    let frame = state.image_space_modifier_frame;
-                    ImageSpaceModifierView {
-                        blur_radius_pixels: frame.blur_radius_pixels,
-                        double_vision_strength: frame.double_vision_strength,
-                        motion_blur_strength: frame.motion_blur_strength,
-                        radial_blur_strength: frame.radial_blur_strength,
-                        radial_blur_ramp_up: frame.radial_blur_ramp_up,
-                        radial_blur_start: frame.radial_blur_start,
-                        radial_blur_ramp_down: frame.radial_blur_ramp_down,
-                        radial_blur_down_start: frame.radial_blur_down_start,
-                        radial_blur_center: frame.radial_blur_center,
-                        saturation: frame.saturation,
-                        brightness: frame.brightness,
-                        contrast: frame.contrast,
-                        tint_color: frame.tint_color,
-                        fade_color: frame.fade_color,
-                    }
-                });
-            let draw_result = ctx.draw_frame(FrameInputs {
-                clear_color,
-                view_proj: &frame.view_proj,
-                draw_commands: &self.draw_commands,
-                lights: &self.gpu_lights,
-                fog_volumes: &self.gpu_fog_volumes,
-                bone_world: &self.bone_world,
-                bind_inverse_pending_uploads: &pending_with_data,
-                materials: self.material_table.materials(),
-                camera_pos: frame.camera_pos,
-                render_origin: frame.render_origin,
-                ambient_color: frame.ambient,
-                fog_color: frame.fog_color,
-                fog_near: frame.fog_near,
-                fog_far: frame.fog_far,
-                fog_extinction_per_meter: frame.fog_medium.extinction_per_meter,
-                fog_single_scatter_albedo: frame.fog_medium.single_scatter_albedo,
-                fog_coverage: frame.fog_medium.coverage,
-                fog_clip: frame.fog_clip,
-                fog_power: frame.fog_power,
-                fog_height_reference: frame.fog_height_reference,
-                ui_texture_handle: ui_tex,
-                sky_params: &frame.sky,
-                dof,
-                frame_time_delta_ms,
-                timings: frame_timings.as_mut(),
-                water_commands: &self.water_commands,
-                underwater: compute_underwater_params(&self.world),
-                image_space_modifier,
-                pose_dirty: self.skin_slot_pool.pose_dirty(),
-            });
-            // #1796 / D6-02 — `draw_frame`'s two early-return `Ok` guards
-            // (empty framebuffers, `ERROR_OUT_OF_DATE_KHR`) are
-            // indistinguishable from a frame that actually reached the skin
-            // dispatch section. The CPU-side pose hash commit already ran
-            // (in `build_render_data`, before `ctx.draw_frame` was called),
-            // so an early return here means that commit needs undoing or the
-            // next frame's dirty gate reads "clean" against a dispatch that
-            // never happened.
-            //
-            // #1791 / D6-01 — the same guards also precede the
-            // `bind_inverses` SSBO upload (draw.rs ~2654-2676), which sits
-            // strictly before the `record_skinned_blas_refit` call that
-            // flips `skin_dispatch_ran` true — so this flag is exactly the
-            // right signal for both bugs. `pending` was already irrevocably
-            // drained from the pool above (before this call), so an early
-            // return here means those first-sight `bind_inverses` were about
-            // to be lost for good: the slot stays resident in
-            // `entity_to_slot` (never re-queued by `allocate`), so the
-            // persistent SSBO region for it is never written, corrupting the
-            // entity's skinning palette for its remaining lifetime in the
-            // cell.
-            //
-            // #2522 / PERF-D6-NEW-01 — `draw_frame` can also return `Err`
-            // (fence wait, command-buffer reset/begin, FSR parameter build —
-            // all execute before `record_skinned_blas_refit`, i.e. while
-            // `skin_dispatch_ran` is still `false`), so this check must run
-            // unconditionally on the `Result`, not only inside the `Ok` arm.
-            if !ctx.skin_dispatch_ran {
-                self.skin_slot_pool.rollback_pending_pose_commits();
-                self.skin_slot_pool
-                    .requeue_pending(std::mem::take(&mut pending_for_requeue));
-            }
-            match draw_result {
-                Ok(needs_recreate) => {
-                    let last_draw_stats = ctx.last_draw_call_stats;
-                    world_resource_set::<DebugStats>(&self.world, |s| {
-                        s.batch_count = last_draw_stats.batch_count;
-                        s.indirect_call_count = last_draw_stats.indirect_call_count;
-                    });
-                    if let Some(ref ft) = frame_timings {
-                        const NS_TO_MS: f32 = 1.0e-6;
-                        let mut cpu_t = self
-                            .world
-                            .resource_mut::<byroredux_core::ecs::CpuFrameTimings>();
-                        cpu_t.fence_wait_ms = ft.fence_wait_ns as f32 * NS_TO_MS;
-                        cpu_t.tlas_build_ms = ft.tlas_build_ns as f32 * NS_TO_MS;
-                        cpu_t.ssbo_build_ms = ft.ssbo_build_ns as f32 * NS_TO_MS;
-                        cpu_t.cmd_record_ms = ft.cmd_record_ns as f32 * NS_TO_MS;
-                        cpu_t.submit_present_ms = ft.submit_present_ns as f32 * NS_TO_MS;
-                        cpu_t.acquire_ms = ft.acquire_ns as f32 * NS_TO_MS;
-                        cpu_t.between_frames_ms = self
-                            .last_redraw_end
-                            .map(|t| t.elapsed().as_nanos() as f32 * NS_TO_MS)
-                            .unwrap_or(0.0);
-                    }
-                    if is_benching {
-                        self.bench_render_ns += render_t0.elapsed().as_nanos() as u64;
-                        if let Some(ft) = frame_timings {
-                            let b = &mut self.bench_frame_timings;
-                            b.fence_wait_ns += ft.fence_wait_ns;
-                            b.tlas_build_ns += ft.tlas_build_ns;
-                            b.ssbo_build_ns += ft.ssbo_build_ns;
-                            b.cmd_record_ns += ft.cmd_record_ns;
-                            b.submit_present_ns += ft.submit_present_ns;
-                        }
-                        if self.bench_start.is_none() {
-                            self.bench_start = Some(Instant::now());
-                        }
-                        self.bench_frames_count += 1;
-                    }
-                    if needs_recreate {
-                        if let Some(ref win) = self.window {
-                            let size = win.inner_size();
-                            if size.width > 0 && size.height > 0 {
-                                if let Err(e) = ctx.recreate_swapchain([size.width, size.height]) {
-                                    log::error!("Swapchain recreate failed: {e:#}");
-                                    event_loop.exit();
-                                }
-                            }
-                        }
-                    }
-                }
-                Err(e) => {
-                    log::error!("Draw failed: {e:#}");
-                    event_loop.exit();
-                }
-            }
-            // Phase 15 — close draw-call bracket. Post-draw
-            // includes the remaining work in this scope plus
-            // the `last_redraw_end` stamp below.
-            rof_draw_call_ns = rof_draw_call_t0.elapsed().as_nanos() as u64;
-        }
-        // Phase 9 — stamp end-of-frame. Next `render_one_frame`
-        // call computes `now() - last_redraw_end` as
-        // `between_frames_ms` (compositor wait + scheduler.run +
-        // about_to_wait host work). Set unconditionally even if
-        // the inner `if let Some(ref mut ctx)` branch was skipped,
-        // so the metric remains continuous across renderer-down
-        // frames.
-        self.last_redraw_end = Some(Instant::now());
-        // Phase 15 — close post-draw bracket and fold the
-        // three-phase split into CpuFrameTimings. atw_post
-        // surfaces render_one_frame's WALL total; this split
-        // shows which phase inside it dominates.
-        const NS_TO_MS: f32 = 1.0e-6;
-        let rof_post_draw_ns = rof_pre_t0
-            .elapsed()
-            .as_nanos()
-            .saturating_sub((rof_pre_draw_ns + rof_draw_call_ns) as u128)
-            as u64;
-        let mut cpu_t = self
-            .world
-            .resource_mut::<byroredux_core::ecs::CpuFrameTimings>();
-        cpu_t.rof_pre_draw_ms = rof_pre_draw_ns as f32 * NS_TO_MS;
-        cpu_t.rof_draw_call_ms = rof_draw_call_ns as f32 * NS_TO_MS;
-        cpu_t.rof_post_draw_ms = rof_post_draw_ns as f32 * NS_TO_MS;
-    }
-}
-
-impl ApplicationHandler for App {
-    fn resumed(&mut self, event_loop: &ActiveEventLoop) {
-        if self.window.is_some() {
-            return;
-        }
-
-        let config = WindowConfig::default();
-
-        let win = match window::create_window(event_loop, &config) {
-            Ok(w) => w,
-            Err(e) => {
-                log::error!("Failed to create window: {e:#}");
-                event_loop.exit();
-                return;
-            }
-        };
-
-        let size = win.inner_size();
-        let (display, window_handle) = match window::raw_handles(&win) {
-            Ok(h) => h,
-            Err(e) => {
-                log::error!("Failed to get raw handles: {e:#}");
-                event_loop.exit();
-                return;
-            }
-        };
-
-        match VulkanContext::new(
-            display,
-            window_handle,
-            [size.width, size.height],
-            self.renderer_config,
-        ) {
-            Ok(ctx) => {
-                // Create screenshot bridge for debug server access.
-                let ss_handle = ctx.screenshot_handle();
-                self.world
-                    .insert_resource(byroredux_core::ecs::ScreenshotBridge {
-                        requested: ss_handle.requested,
-                        result: ss_handle.result,
-                        // #1006 — owner-tagged claim so the CLI
-                        // `--screenshot` deadline loop and the
-                        // debug-server `DebugRequest::Screenshot`
-                        // can't race on a single result slot.
-                        // Starts idle (SCREENSHOT_OWNER_NONE).
-                        owner: std::sync::Arc::new(std::sync::atomic::AtomicU8::new(
-                            byroredux_core::ecs::resources::SCREENSHOT_OWNER_NONE,
-                        )),
-                        // #1603 — shared capture generation; the renderer
-                        // gates each readback's publish on it so a
-                        // cancelled-then-resumed straggler is discarded.
-                        generation: ss_handle.generation,
-                    });
-
-                // Expose the GPU allocator to the ECS so the
-                // `mem.frag` console command can compute a live
-                // fragmentation report on demand. Newtype wrapper
-                // dodges the orphan rule on `Resource`. See #503.
-                if let Some(ref alloc) = ctx.allocator {
-                    self.world.insert_resource(
-                        byroredux_renderer::vulkan::allocator::AllocatorResource(alloc.clone()),
-                    );
-                }
-
-                // Cache the VRAM budget once — heap sizes are immutable
-                // after device pick. Read by `metrics_sample_system` to
-                // compute the `used / budget` ratio without a per-frame
-                // `vkGetPhysicalDeviceMemoryProperties` round trip.
-                self.world.insert_resource(
-                    byroredux_renderer::vulkan::allocator::GpuMemoryBudget::sample(
-                        &ctx.instance,
-                        ctx.physical_device,
-                    ),
-                );
-
-                // Phase 4 of the debug-UI plan — initialise the
-                // egui overlay before the first frame.
-                let mut ctx = ctx;
-                if let Err(e) =
-                    ctx.init_egui(byroredux_renderer::vulkan::sync::MAX_FRAMES_IN_FLIGHT)
-                {
-                    log::warn!("debug-UI overlay init failed: {e:#}");
-                }
-                let debug_ui_state = byroredux_debug_ui::DebugUiState::new(event_loop, &win);
-                debug_ui_state.sync_registered_settings(&self.world.resource::<SettingsRegistry>());
-                self.debug_ui = Some(debug_ui_state);
-
-                self.renderer = Some(ctx);
-                self.window = Some(win);
-                self.last_frame = Instant::now();
-                self.setup_scene();
-                // Preserve the scene/CLI-authored camera before the startup
-                // scheduler's character-follow system overwrites it. The
-                // frame loop will derive and reapply the requested bench path
-                // from this seed on every rendered frame.
-                self.seed_bench_camera_origin();
-                // M41.0 Phase 1b.x — Prime the scene's transform state
-                // BEFORE the event loop starts.
-                self.scheduler.run(&self.world, 0.0);
-                self.renderer.as_ref().unwrap().log_memory_usage();
-                log::info!("Engine ready — entering game loop");
-            }
-            Err(e) => {
-                log::error!("Vulkan init failed: {e:#}");
-                event_loop.exit();
-            }
-        }
-    }
-
-    fn window_event(
-        &mut self,
-        event_loop: &ActiveEventLoop,
-        _window_id: WindowId,
-        event: WindowEvent,
-    ) {
-        // Debug-UI event forwarding — egui sees every WindowEvent
-        // before the camera input layer. When the overlay is
-        // visible AND egui claims to have consumed the event (e.g.
-        // a click inside an egui window, a keypress targeting a
-        // text field), the rest of the dispatch is skipped so the
-        // fly camera doesn't move with the cursor that's busy
-        // dragging an egui slider. CloseRequested + Resized always
-        // run their normal handlers — egui doesn't care about
-        // those.
-        let egui_consumed = if let (Some(ref mut state), Some(win)) =
-            (self.debug_ui.as_mut(), self.window.as_ref())
-        {
-            state.on_window_event(win, &event).consumed
-        } else {
-            false
-        };
-        if egui_consumed && !matches!(event, WindowEvent::CloseRequested | WindowEvent::Resized(_))
-        {
-            return;
-        }
-        if self.route_scaleform_window_event(&event) {
-            return;
-        }
-
-        match event {
-            WindowEvent::CloseRequested => {
-                log::info!("Close requested — shutting down");
-                // M40 streaming cleanup: every per-cell streamed load owns
-                // GPU resources (BLAS / mesh buffers / texture refcounts)
-                // released only via `cell_loader::unload_cell`. Without
-                // this sweep the allocator finds 1 dangling ref per
-                // (BLAS + mesh + texture) per loaded cell at ctx
-                // destruction time, triggers the
-                // "leaking allocator to avoid use-after-free" path, and
-                // SIGSEGVs as the orphaned device handles get reaped.
-                if let (Some(ref mut state), Some(ref mut ctx)) =
-                    (self.streaming.as_mut(), self.renderer.as_mut())
-                {
-                    let cells: Vec<_> = state.loaded.drain().collect();
-                    log::info!(
-                        "Streaming shutdown: unloading {} streamed cells before ctx destroy",
-                        cells.len()
-                    );
-                    for ((_gx, _gy), slot) in cells {
-                        cell_loader::unload_cell(&mut self.world, ctx, slot.cell_root);
-                    }
-                    // #732 / LIFE-H2 — `unload_cell` queues per-cell
-                    // BLAS/mesh/texture into the renderer's deferred-
-                    // destroy lists with a `MAX_FRAMES_IN_FLIGHT`
-                    // countdown. The countdown only ticks inside
-                    // `draw_frame`, but the window-close path goes
-                    // straight from the unload sweep to `ctx Drop` with
-                    // no intervening render frames. `Drop`'s in-block
-                    // drain catches them eventually, but doing the
-                    // drain explicitly here releases the per-queue
-                    // entries' `Arc<Mutex<Allocator>>` clones before we
-                    // even start tearing down the context — keeps the
-                    // shutdown ordering visible at the call site rather
-                    // than buried inside `VulkanContext::Drop`.
-                    ctx.flush_pending_destroys();
-                }
-                // Drop the streaming state explicitly — joins the
-                // worker thread cleanly before we tear down the GPU.
-                // Without this, the worker could still be holding
-                // `Arc<TextureProvider>` file handles that the
-                // allocator's leak path would observe as outstanding
-                // refs.
-                //
-                // #856 / C6-NEW-03 — pre-fix this was a bare
-                // `self.streaming.take()` which detached the worker
-                // (the `JoinHandle` was dropped on the same line via
-                // `WorldStreamingState` Drop). `shutdown` drops
-                // `request_tx` first, then joins with a 1-second
-                // bound so a slow `BsaArchive::extract()` can't
-                // block process teardown.
-                if let Some(mut state) = self.streaming.take() {
-                    state.shutdown(std::time::Duration::from_secs(1));
-                    // `state` goes out of scope here. Drop runs but
-                    // `shutdown` already took `worker` and `request_tx`
-                    // so it short-circuits — the join is not repeated.
-                }
-                // Release the ECS clone of the GPU allocator before
-                // dropping the renderer.  VulkanContext::Drop calls
-                // Arc::try_unwrap on the allocator; if AllocatorResource
-                // is still in the world it holds an extra strong-count
-                // that makes try_unwrap fail, triggering the
-                // device+surface+instance leak path (#1406 / MEM-03).
-                self.world
-                    .remove_resource::<byroredux_renderer::vulkan::allocator::AllocatorResource>();
-                self.renderer.take();
-                self.window.take();
-                event_loop.exit();
-            }
-            WindowEvent::Resized(size) => {
-                if let Some(ref mut ctx) = self.renderer {
-                    if size.width > 0 && size.height > 0 {
-                        if let Err(e) = ctx.recreate_swapchain([size.width, size.height]) {
-                            log::error!("Swapchain recreate failed: {e:#}");
-                            event_loop.exit();
-                        }
-                        // Update camera aspect ratio.
-                        if let Some(active) = self.world.try_resource::<ActiveCamera>() {
-                            let cam_entity = active.0;
-                            drop(active);
-                            if let Some(mut q) = self.world.query_mut::<Camera>() {
-                                if let Some(cam) = q.get_mut(cam_entity) {
-                                    cam.aspect = size.width as f32 / size.height as f32;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            WindowEvent::RedrawRequested => {
-                // Phase 14 — render is now driven by `about_to_wait`,
-                // not by the compositor's `RedrawRequested` event.
-                // The OS still fires this on window expose / resize /
-                // first paint, but we don't render here — the next
-                // `about_to_wait` tick will do the work and present
-                // the new frame. Keeping the arm empty (not removed)
-                // so the match stays exhaustive against the existing
-                // dummy match scope; the body is intentionally bare.
-            }
-            WindowEvent::KeyboardInput { event, .. } => {
-                if let PhysicalKey::Code(code) = event.physical_key {
-                    let mut input = self.world.resource_mut::<InputState>();
-                    match event.state {
-                        ElementState::Pressed => {
-                            // Escape toggles mouse capture.
-                            if code == KeyCode::Escape && !event.repeat {
-                                let captured = !input.mouse_captured;
-                                input.mouse_captured = captured;
-                                drop(input);
-                                if let Some(ref win) = self.window {
-                                    if captured {
-                                        let _ =
-                                            win.set_cursor_grab(CursorGrabMode::Confined).or_else(
-                                                |_| win.set_cursor_grab(CursorGrabMode::Locked),
-                                            );
-                                        win.set_cursor_visible(false);
-                                    } else {
-                                        let _ = win.set_cursor_grab(CursorGrabMode::None);
-                                        win.set_cursor_visible(true);
-                                    }
-                                }
-                            } else if code == KeyCode::KeyF && !event.repeat {
-                                // M28.5 follow-up — Walk ↔ Fly mode toggle.
-                                // Temporary debug binding until an in-engine
-                                // console (byro-dbg embed) is available. Models
-                                // Bethesda's `tcl` (toggle collision) command:
-                                // - Fly → Character: snap the character body to
-                                //   the camera's current world position (so the
-                                //   player "lands" wherever the freeflight cam
-                                //   was looking from). The character_controller
-                                //   then takes over from there.
-                                // - Character → Fly: no-op on positions —
-                                //   `camera_follow_system` had been writing the
-                                //   active camera at `body_pos + eye_height`
-                                //   anyway, so the fly cam takes over from the
-                                //   same place. The character body stays alive
-                                //   but `character_controller_system` early-
-                                //   returns on FlyCam mode, so it freezes in
-                                //   place until the user toggles back.
-                                drop(input);
-                                toggle_player_mode(&mut self.world);
-                            } else if code == KeyCode::F3 && !event.repeat {
-                                // Phase 4 of the debug-UI plan — F3
-                                // toggles the egui overlay. Doesn't
-                                // touch InputState so the camera
-                                // input layer is uninterrupted (the
-                                // overlay's own egui-winit handler
-                                // is the source of truth for any
-                                // mouse / keyboard egui needs).
-                                drop(input);
-                                if let Some(ref mut ui) = self.debug_ui {
-                                    ui.toggle();
-                                    // #2166 — the Metrics panel is the only
-                                    // consumer of the scheduler's per-system
-                                    // wall times, and the resource's presence
-                                    // is what arms that tracker. Insert it the
-                                    // first time the overlay opens rather than
-                                    // at world setup, so a run that never opens
-                                    // the overlay never pays for it. Left in
-                                    // place once armed: a toggle-off would
-                                    // otherwise blank the panel's history on
-                                    // every reopen, and the idle cost of the
-                                    // empty resource is a single probe.
-                                    if ui.visible
-                                        && self
-                                            .world
-                                            .try_resource::<byroredux_core::ecs::SchedulerSystemTimings>()
-                                            .is_none()
-                                    {
-                                        self.world.insert_resource(
-                                            byroredux_core::ecs::SchedulerSystemTimings::default(),
-                                        );
-                                    }
-                                }
-                            } else {
-                                input.keys_held.insert(code);
-                            }
-                        }
-                        ElementState::Released => {
-                            input.keys_held.remove(&code);
-                        }
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn device_event(
-        &mut self,
-        _event_loop: &ActiveEventLoop,
-        _device_id: DeviceId,
-        event: DeviceEvent,
-    ) {
-        if let DeviceEvent::MouseMotion { delta } = event {
-            let ui_focused = self
-                .ui_manager
-                .as_ref()
-                .is_some_and(UiManager::has_input_focus);
-            if ui_focused {
-                self.release_world_input_for_ui();
-                return;
-            }
-            let looking_enabled = self
-                .world
-                .try_resource::<byroredux_scripting::PlayerControlState>()
-                .map(|controls| controls.looking_enabled)
-                .unwrap_or(true);
-            if !looking_enabled {
-                return;
-            }
-            let mut input = self.world.resource_mut::<InputState>();
-            if input.mouse_captured {
-                let sensitivity = input.look_sensitivity;
-                input.yaw -= delta.0 as f32 * sensitivity;
-                input.pitch -= delta.1 as f32 * sensitivity;
-                // Clamp pitch to avoid flipping.
-                input.pitch = input.pitch.clamp(
-                    -std::f32::consts::FRAC_PI_2 + 0.01,
-                    std::f32::consts::FRAC_PI_2 - 0.01,
-                );
-            }
-        }
-    }
-
-    fn about_to_wait(&mut self, event_loop: &ActiveEventLoop) {
-        // Menu focus can change through engine code without a corresponding
-        // winit event. Enforce modal ownership before the scheduler reads
-        // InputState so a held movement key cannot leak for one frame.
-        if self
-            .ui_manager
-            .as_ref()
-            .is_some_and(UiManager::has_input_focus)
-        {
-            self.release_world_input_for_ui();
-        }
-
-        let atw_pre_t0 = Instant::now();
-        let now = atw_pre_t0;
-        // Finite benchmarks resolve one named mode before the event loop
-        // starts, so delta-time cannot drift independently of camera policy.
-        // Outside a finite bench, retain BYROREDUX_FIXED_DT as a diagnostic
-        // override for tools that do not emit benchmark conclusions.
-        let wall_dt = now.duration_since(self.last_frame).as_secs_f32();
-        let dt = if crate::bench::harness_active(self.bench_summary_printed) {
-            self.bench_mode.map_or_else(
-                || {
-                    // Preserve the environment override for non-benchmark
-                    // tools. Finite benches resolve it once into a named mode
-                    // in boot.rs.
-                    std::env::var("BYROREDUX_FIXED_DT")
-                        .ok()
-                        .and_then(|s| s.parse::<f32>().ok())
-                        .unwrap_or(wall_dt)
-                },
-                |mode| mode.delta_time(wall_dt),
-            )
-        } else {
-            // A held session becomes interactive after its finite benchmark:
-            // fixed dt (including a legacy environment override resolved by
-            // the bench) must no longer suppress walk/fly movement.
-            wall_dt
-        };
-        self.last_frame = now;
-
-        // Update time resources.
-        world_resource_set::<DeltaTime>(&self.world, |r| r.0 = dt);
-        world_resource_set::<TotalTime>(&self.world, |r| r.0 += dt);
-
-        // Update debug stats.
-        //
-        // #637 / FNV-D5-02 — `mesh_count` / `texture_count` are
-        // registry-wide and don't drop on cell unload. The new
-        // `meshes_in_use` / `textures_in_use` counts walk the ECS
-        // `MeshHandle` / `TextureHandle` queries and dedupe non-zero
-        // handles, so a regression that retains a registry entry past
-        // the last live consumer shows up as `registry > in_use`. Done
-        // in two scopes because the queries need an immutable world
-        // borrow that can't coexist with `resource_mut::<DebugStats>`.
-        //
-        // PERF-D1-NEW-01 / #1801 — this walk used to run unconditionally
-        // every frame, but both consumers (the `stats` console command
-        // and the debug-server entity evaluator) are on-demand, not
-        // per-frame; `log_stats_system` doesn't print these fields
-        // either. Throttled to the same once-per-wall-clock-second
-        // boundary `log_stats_system` already uses for its own summary
-        // line, so a console/debug-server read is at most ~1 second
-        // stale — indistinguishable from before for a human operator,
-        // for a cost paid once/second instead of every frame.
-        let total = self.world.resource::<byroredux_core::ecs::TotalTime>().0;
-        let should_refresh_handle_counts = crate::systems::crosses_one_second_boundary(total, dt);
-        if should_refresh_handle_counts {
-            // #1584 — reuse persistent scratch sets (clear() keeps the
-            // allocation, drops the contents) so this dedup walk does
-            // zero steady-state heap allocations.
-            self.in_use_mesh_scratch.clear();
-            if let Some(q) = self.world.query::<byroredux_core::ecs::MeshHandle>() {
-                for (_, h) in q.iter() {
-                    if h.0 != 0 {
-                        self.in_use_mesh_scratch.insert(h.0);
-                    }
-                }
-            }
-            self.in_use_tex_scratch.clear();
-            if let Some(q) = self.world.query::<byroredux_core::ecs::TextureHandle>() {
-                for (_, h) in q.iter() {
-                    if h.0 != 0 {
-                        self.in_use_tex_scratch.insert(h.0);
-                    }
-                }
-            }
-        }
-        {
-            let mut stats = self.world.resource_mut::<DebugStats>();
-            stats.push_frame_time(dt);
-            stats.entity_count = self.world.next_entity_id();
-            // Off-cadence frames keep the previous values (still fresh to
-            // within ~1 second) rather than stale-to-zero.
-            if should_refresh_handle_counts {
-                stats.meshes_in_use = self.in_use_mesh_scratch.len() as u32;
-                stats.textures_in_use = self.in_use_tex_scratch.len() as u32;
-            }
-            if let Some(ref ctx) = self.renderer {
-                stats.mesh_count = ctx.mesh_registry.len() as u32;
-                stats.texture_count = ctx.texture_registry.len() as u32;
-            }
-            // #1284 — mirror SkinSlotPool telemetry into DebugStats so
-            // `log_stats_system` (ECS, no App access) can surface it.
-            stats.skin_pool_live = self.skin_slot_pool.live_slot_count();
-            stats.skin_pool_max = self.skin_slot_pool.max_slot();
-            stats.skin_pool_overflow_attempts = self.skin_slot_pool.overflow_attempt_count();
-        }
-
-        // Refresh renderer-side scratch-Vec telemetry (R6). Reuses the
-        // resource's `rows` Vec so this is amortized to ~zero allocs
-        // after the first frame; capacity stabilises at the count of
-        // declared scratches in `VulkanContext::fill_scratch_telemetry`.
-        if let Some(ref ctx) = self.renderer {
-            let mut tlm = self.world.resource_mut::<ScratchTelemetry>();
-            ctx.fill_scratch_telemetry(&mut tlm.rows);
-        }
-
-        // EX-05 / #2736 — mirror the pre-tonemap non-finite pixel counters so
-        // the console and bench summary can read them without renderer access.
-        if let Some(ref ctx) = self.renderer {
-            let ((last_rgb, last_alpha), (total_rgb, total_alpha)) = ctx.image_health();
-            let mut health = self
-                .world
-                .resource_mut::<byroredux_core::ecs::ImageHealth>();
-            health.last_non_finite_rgb = last_rgb;
-            health.last_non_finite_alpha = last_alpha;
-            health.total_non_finite_rgb = total_rgb;
-            health.total_non_finite_alpha = total_alpha;
-        }
-
-        // EX-08 / #2374 — refresh the cross-subsystem ownership snapshot on the
-        // same throttled cadence as the handle counts above. It reuses the
-        // scratch sets those already populated rather than re-walking the ECS,
-        // so the added per-sample cost is a handful of `len()` calls; off-
-        // cadence frames keep the previous sample.
-        if should_refresh_handle_counts {
-            let snapshot = crate::ownership_sample::sample_all(
-                &self.world,
-                self.renderer.as_ref(),
-                self.in_use_mesh_scratch.len(),
-                self.in_use_tex_scratch.len(),
-            );
-            self.world
-                .resource_mut::<byroredux_core::ecs::OwnershipTelemetry>()
-                .current = snapshot;
-        }
-
-        // Refresh skinned-BLAS coverage stats — captures last frame's
-        // dispatches / first-sight / refit counters from the renderer
-        // so `skin.coverage` reflects the just-drawn frame. Mirrors the
-        // scratch-telemetry pattern; the `failed_entity_ids` Vec is
-        // bounded to 16 entries inside `fill_skin_coverage_stats`.
-        if let Some(ref ctx) = self.renderer {
-            let mut cov = self.world.resource_mut::<SkinCoverageStats>();
-            ctx.fill_skin_coverage_stats(&mut cov);
-        }
-
-        // Refresh the upscaler line `ctx.upscaler` prints — the FSR provider
-        // version and the SDK's own GPU reservation, which live outside
-        // `gpu-allocator` and so never appear in `ctx.memory`.
-        if let Some(ref ctx) = self.renderer {
-            let mut telemetry = self
-                .world
-                .resource_mut::<byroredux_core::ecs::UpscalerTelemetry>();
-            ctx.fill_upscaler_telemetry(&mut telemetry);
-        }
-
-        // Select the deterministic bench-camera pose before the scheduler so
-        // camera-dependent systems observe this frame's requested path.
-        self.step_bench_camera();
-
-        // End of pre-scheduler phase (Phase 10 bracket).
-        let atw_pre_ns = atw_pre_t0.elapsed().as_nanos() as u64;
-
-        // Run all systems.
-        let systems_t0 = Instant::now();
-        self.scheduler.run(&self.world, dt);
-        let atw_scheduler_ns = systems_t0.elapsed().as_nanos() as u64;
-        if self.bench_frames_target.is_some() && self.renderer.is_some() {
-            self.bench_systems_ns += atw_scheduler_ns;
-            self.bench_systems_ticks += 1;
-        }
-
-        // Post-scheduler phase starts here (Phase 10 bracket).
-        let atw_post_t0 = Instant::now();
-
-        // Character-mode camera sync runs inside the scheduler and otherwise
-        // replaces an explicit --camera-pos / --bench-camera pose with the
-        // player capsule's eye transform. Restore the selected bench pose
-        // before both streaming and rendering. This is a no-op outside a
-        // deterministic bench.
-        self.restore_bench_camera_pose();
-
-        // World cell streaming (M40 Phase 1a). Runs after the
-        // scheduler so the scheduler-driven `fly_camera_system` has
-        // already published the player's current Transform translation
-        // for this frame. No-ops outside `--esm + --grid` exterior
-        // mode and when the player hasn't crossed a boundary.
-        self.step_streaming();
-
-        // Debug-UI load queue (Phase 2 of the debug-UI plan). Drains
-        // the `PendingDebugLoadSlot` populated by the debug-server's
-        // `LoadNif` / `LoadInteriorCell` / `LoadExteriorCell`
-        // handlers. Sequenced BEFORE `step_cell_transition` so a
-        // debug load that arrives the same frame as a queued
-        // `door.teleport` doesn't trample the transition's mid-load
-        // state.
-        self.step_debug_loads();
-        self.step_upscaler_switch(event_loop);
-
-        // M45.1 refinement — snapshot the player/camera pose now that the
-        // scheduler's camera systems have published this frame's Transform,
-        // so a `save` triggered this frame records where the player stands.
-        crate::save_io::capture_player_pose(&self.world);
-
-        // M45.1 — live save-load: reload the saved cell + overlay saved
-        // form-id-keyed deltas. Runs alongside the other deferred drains,
-        // no-op when no `load` is queued.
-        self.step_save_loads();
-
-        // Cell-transition dispatch (M40 Phase 2 Stage 3). Drains the
-        // `PendingCellTransitionSlot` posted by `door.teleport`
-        // (and future F-key activate) and dispatches the orchestrator.
-        // No-op when the slot is `None` — the common per-frame case.
-        self.step_cell_transition();
-
-        // Update window title with stats (throttled: every 16 frames ≈ 4×/sec at 60fps).
-        let config_debug = self.world.resource::<EngineConfig>().debug_logging;
-        if config_debug {
-            let stats = self.world.resource::<DebugStats>();
-            if stats.frame_index().is_multiple_of(16) {
-                if let Some(ref win) = self.window {
-                    // #1258 — `{}/{}b/{}c draws` = input DrawCommands /
-                    // post-merge batches / actual GPU calls.
-                    win.set_title(&format!(
-                        "ByroRedux | {:.0} FPS | {:.1}ms | {} entities | {} meshes | {} textures | {}/{}b/{}c draws",
-                        stats.avg_fps(), stats.frame_time_ms,
-                        stats.entity_count, stats.mesh_count, stats.texture_count,
-                        stats.draw_command_count, stats.batch_count, stats.indirect_call_count,
-                    ));
-                }
-            }
-        }
-
-        // Phase 14 — drive rendering directly from `about_to_wait`
-        // instead of `win.request_redraw()` → wait for
-        // compositor frame callback → `WindowEvent::RedrawRequested`.
-        // On Wayland + winit 0.30 the indirection costs ~54 ms per
-        // frame at the compositor's pace. Drawing here uncaps the
-        // loop; MAILBOX present mode still vsyncs the actual
-        // presentation but `between_frames` drops to the
-        // GPU+CPU-bound minimum.
-        if self.window.is_some() && self.renderer.is_some() {
-            self.render_one_frame(event_loop);
-        }
-        if self.bench_frames_target.is_some() {
-            let samples = self.bench_cpu_frame_ms.len() as u32;
-            if samples < self.bench_frames_count {
-                self.bench_cpu_frame_ms
-                    .push(atw_pre_t0.elapsed().as_secs_f64() * 1000.0);
-            }
-        }
-
-        // --bench-frames: once we've rendered the target number of
-        // frames, emit a single machine-readable summary line and exit.
-        // The renderer must be up (bench counts start after the first
-        // real frame); a `--bench-frames N` that never renders (window
-        // creation fails, etc.) does nothing here.
-        if let Some(target) = self.bench_frames_target {
-            if self.renderer.is_some() {
-                // Guard: under `--bench-hold` we re-enter this branch on
-                // every `about_to_wait` tick once the bench window has
-                // closed; without the `bench_summary_printed` flag the
-                // summary would dump per-tick and the screenshot path
-                // would re-fire forever.
-                // Both streaming paths run on the logical clock, so both must
-                // ask the clock (not the rendered-frame count) whether they are
-                // done — otherwise a soak would end mid-boundary and report the
-                // in-flight cell as a leaked owner.
-                let uses_streaming_clock = matches!(
-                    self.bench_camera,
-                    Some(crate::bench_camera::BenchCameraPath::GridCross)
-                        | Some(crate::bench_camera::BenchCameraPath::GridSoak)
-                );
-                let path_complete = if uses_streaming_clock {
-                    let boundary_in_progress = self
-                        .streaming
-                        .as_ref()
-                        .is_some_and(|state| state.telemetry.boundary_in_progress());
-                    crate::bench_camera::grid_cross_complete(
-                        self.bench_camera_path_frame,
-                        target,
-                        boundary_in_progress,
-                    )
-                } else {
-                    self.bench_frames_count >= target
-                };
-                if path_complete && !self.bench_summary_printed {
-                    // Capture the renderer-facing state only after the timed
-                    // window has closed. Hashing a large bone palette can be
-                    // measurable CPU work, so it must not feed back into the
-                    // frame time being reported.
-                    let scene_state = crate::bench::capture_scene_state(
-                        &self.world,
-                        &self.draw_commands,
-                        &self.water_commands,
-                        &self.gpu_lights,
-                        &self.gpu_fog_volumes,
-                        &self.bone_world,
-                    );
-                    let bench_mode = self
-                        .bench_mode
-                        .expect("every finite benchmark resolves a named mode");
-                    let bench_camera = self
-                        .bench_camera
-                        .map_or_else(|| "free".to_owned(), |path| path.to_string());
-                    let stats = self.world.resource::<DebugStats>();
-                    let elapsed_secs = self
-                        .bench_start
-                        .map(|t| t.elapsed().as_secs_f64())
-                        .unwrap_or(1.0);
-                    let wall_fps = self.bench_frames_count as f64 / elapsed_secs;
-                    let wall_ms = elapsed_secs * 1000.0 / self.bench_frames_count as f64;
-                    let [frame_p50_ms, frame_p95_ms, frame_max_ms] =
-                        bench_frame_distribution(&self.bench_cpu_frame_ms);
-                    let n = self.bench_frames_count as f64;
-                    let ticks_per_frame = self.bench_systems_ticks as f64 / n;
-                    let systems_ms = if self.bench_systems_ticks > 0 {
-                        self.bench_systems_ns as f64 / self.bench_systems_ticks as f64 / 1e6
-                    } else {
-                        0.0
-                    };
-                    let brd_ms = self.bench_build_render_ns as f64 / n / 1e6;
-                    let ui_ms = self.bench_ui_ns as f64 / n / 1e6;
-                    let draw_ms = self.bench_render_ns as f64 / n / 1e6;
-                    let ft = &self.bench_frame_timings;
-                    let fence_ms = ft.fence_wait_ns as f64 / n / 1e6;
-                    let tlas_ms = ft.tlas_build_ns as f64 / n / 1e6;
-                    let ssbo_ms = ft.ssbo_build_ns as f64 / n / 1e6;
-                    let cmd_ms = ft.cmd_record_ns as f64 / n / 1e6;
-                    let submit_ms = ft.submit_present_ns as f64 / n / 1e6;
-                    let accounted = systems_ms * ticks_per_frame + brd_ms + ui_ms + draw_ms;
-                    let unaccounted_ms = (wall_ms - accounted).max(0.0);
-                    // #1194 — per-pass GPU timer snapshot. The
-                    // SkinCoverageStats resource is filled at the end
-                    // of every `draw_frame`; values here are from the
-                    // last completed frame and represent one
-                    // `MAX_FRAMES_IN_FLIGHT` cycle of pipeline lag.
-                    // Reads 0.0 across the board when the driver
-                    // lacks `timestampComputeAndGraphics` or no
-                    // skinned/TAA work fired on the snapshot frame.
-                    // Surfaces `gpu_skin_disp` / `gpu_blas_refit` /
-                    // `gpu_taa` so PERF-DIM7-01/-02/-03 (#1195/#1196/
-                    // #1197) can measure rather than guess.
-                    // Full per-pass GPU breakdown. The FSR benchmark matrix
-                    // (execution phase 7) needs the render-resolution passes
-                    // and the output-resolution ones separable, because only
-                    // the former shrink with a preset — reporting a frame-time
-                    // win without netting out presentation and the upscale
-                    // dispatch would overstate what a player actually gets.
-                    let gpu = self
-                        .world
-                        .try_resource::<byroredux_core::ecs::SkinCoverageStats>()
-                        .map(|s| {
-                            // Copy just the timing fields out from under the
-                            // resource read guard; the resource itself carries
-                            // non-Copy diagnostic state we do not need here.
-                            [
-                                s.gpu_skin_dispatch_ms,
-                                s.gpu_skin_blas_refit_ms,
-                                s.gpu_taa_ms,
-                                s.gpu_upscale_ms,
-                                s.gpu_main_render_ms,
-                                s.gpu_svgf_ms,
-                                s.gpu_composite_ms,
-                                s.gpu_ssao_ms,
-                                s.gpu_bloom_ms,
-                                s.gpu_volumetrics_ms,
-                                s.gpu_cluster_cull_ms,
-                                s.gpu_presentation_ms,
-                            ]
-                        })
-                        .unwrap_or([0.0; 12]);
-                    println!(
-                        "bench: mode={} gate={} dt={} camera={} frames={} \
-                         wall_fps={:.1} wall_ms={:.2} \
-                         frame_p50_ms={:.2} frame_p95_ms={:.2} frame_max_ms={:.2} \
-                         brd_ms={:.2} ui_ms={:.2} draw_ms={:.2} \
-                         [fence={:.2} tlas_ms={:.2} ssbo={:.2} cmd={:.2} submit={:.2}] \
-                         [gpu_skin_disp={:.3} gpu_blas_refit={:.3} gpu_taa={:.3} \
-                         gpu_upscale={:.3} gpu_main_render={:.3} gpu_svgf={:.3} \
-                         gpu_composite={:.3} gpu_ssao={:.3} gpu_bloom={:.3} \
-                         gpu_volumetrics={:.3} gpu_cluster_cull={:.3} \
-                         gpu_presentation={:.3}] \
-                         systems_ms={:.2} ticks_per_frame={:.1} unaccounted_ms={:.2} \
-                         camera_pos={:.3},{:.3},{:.3} camera_forward={:.6},{:.6},{:.6} \
-                         sim_time_s={:.6} entities={} meshes={} textures={} \
-                         draws={}/{}b/{}c lights={} tlas={} state_hash={:016x}",
-                        bench_mode,
-                        bench_mode.gate_label(),
-                        bench_mode.dt_label(),
-                        bench_camera,
-                        self.bench_frames_count,
-                        wall_fps,
-                        wall_ms,
-                        frame_p50_ms,
-                        frame_p95_ms,
-                        frame_max_ms,
-                        brd_ms,
-                        ui_ms,
-                        draw_ms,
-                        fence_ms,
-                        tlas_ms,
-                        ssbo_ms,
-                        cmd_ms,
-                        submit_ms,
-                        gpu[0],
-                        gpu[1],
-                        gpu[2],
-                        gpu[3],
-                        gpu[4],
-                        gpu[5],
-                        gpu[6],
-                        gpu[7],
-                        gpu[8],
-                        gpu[9],
-                        gpu[10],
-                        gpu[11],
-                        systems_ms,
-                        ticks_per_frame,
-                        unaccounted_ms,
-                        scene_state.camera_position[0],
-                        scene_state.camera_position[1],
-                        scene_state.camera_position[2],
-                        scene_state.camera_forward[0],
-                        scene_state.camera_forward[1],
-                        scene_state.camera_forward[2],
-                        scene_state.simulated_time_s,
-                        scene_state.entities,
-                        stats.mesh_count,
-                        stats.texture_count,
-                        // #1258 — `draws=N/Mb/Kc` = N input DrawCommands
-                        // / M post-merge batches / K actual GPU calls.
-                        // Pre-fix this was a single `draws=N` that
-                        // looked like a GPU call count but was actually
-                        // the input. Format change preserves the
-                        // existing first number for audit comparability.
-                        scene_state.draws,
-                        stats.batch_count,
-                        stats.indirect_call_count,
-                        scene_state.lights,
-                        scene_state.tlas_eligible,
-                        scene_state.state_hash,
-                    );
-                    drop(stats);
-                    if let Some(streaming) = self.streaming.as_ref() {
-                        println!("{}", streaming.telemetry.bench_line());
-                    }
-
-                    // --screenshot: queue a capture request and defer
-                    // the event-loop exit until the PNG lands (or the
-                    // frame-budget elapses). The screenshot flow takes
-                    // 2+ frames: frame N kicks the staging copy, N+1
-                    // encodes the PNG. We re-enter this branch up to
-                    // SCREENSHOT_DEADLINE_FRAMES times before giving
-                    // up.
-                    if let Some(path) = self.screenshot_path.clone() {
-                        if !self.screenshot_requested {
-                            if let Some(bridge) = self
-                                .world
-                                .try_resource::<byroredux_core::ecs::ScreenshotBridge>()
-                            {
-                                // #1006 — claim ownership atomically.
-                                // If the debug-server already holds
-                                // the bridge (rare: byro-dbg attached
-                                // before the CLI's first frame issues
-                                // its screenshot command), bail with a
-                                // clear error so the user knows the
-                                // collision happened instead of silently
-                                // racing for the result slot.
-                                if !bridge
-                                    .try_claim(byroredux_core::ecs::resources::SCREENSHOT_OWNER_CLI)
-                                {
-                                    eprintln!(
-                                        "screenshot: bridge already claimed (debug-server owns it) — skipping CLI capture"
-                                    );
-                                    self.screenshot_path = None;
-                                } else {
-                                    drop(bridge);
-                                    self.screenshot_requested = true;
-                                    self.screenshot_deadline_frames = 60;
-                                    return; // keep running frames
-                                }
-                            }
-                        }
-
-                        // Poll the result slot until the PNG arrives.
-                        // Owner-gated take so a debug-server screenshot
-                        // racing past the CLI claim can't steal our bytes.
-                        let maybe_bytes = self
-                            .world
-                            .try_resource::<byroredux_core::ecs::ScreenshotBridge>()
-                            .and_then(|b| {
-                                b.take_result_for(
-                                    byroredux_core::ecs::resources::SCREENSHOT_OWNER_CLI,
-                                )
-                            });
-                        if let Some(bytes) = maybe_bytes {
-                            match std::fs::write(&path, &bytes) {
-                                Ok(()) => {
-                                    println!("screenshot: wrote {} bytes to {}", bytes.len(), path)
-                                }
-                                Err(e) => eprintln!("screenshot: failed to write {}: {e}", path),
-                            }
-                        } else if self.screenshot_deadline_frames > 0 {
-                            self.screenshot_deadline_frames -= 1;
-                            return; // keep pumping
-                        } else {
-                            eprintln!("screenshot: timed out waiting for PNG result",);
-                        }
-                    }
-
-                    // Latch the summary-once invariant for `--bench-hold`
-                    // and only exit when the caller hasn't asked to hold
-                    // the engine open. Under hold, the next about_to_wait
-                    // ticks render normal frames + service the debug
-                    // server (port 9876 by default) so `byro-dbg` can
-                    // attach and run console commands against the loaded
-                    // scene. See `--bench-hold` in main() and the FNV-D5
-                    // audit's coverage gap.
-                    self.bench_summary_printed = true;
-                    if !self.bench_hold {
-                        event_loop.exit();
-                    } else {
-                        eprintln!(
-                            "bench-hold: engine held open in live interactive mode — \
-                             attach via `cargo run -p byro-dbg` \
-                             (port {}). Ctrl+C / window close to exit.",
-                            std::env::var("BYRO_DEBUG_PORT").unwrap_or_else(|_| "9876".to_string()),
-                        );
-                    }
-                }
-            }
-        }
-
-        // Phase 10 — write the about_to_wait phase timings into
-        // `CpuFrameTimings` so the egui Metrics panel can show
-        // where the 501 ms `between_frames` gap (Phase 9) is
-        // actually spent inside this handler. Pre / scheduler /
-        // post split lets the operator localize without
-        // per-system instrumentation.
-        const NS_TO_MS: f32 = 1.0e-6;
-        let atw_post_ns = atw_post_t0.elapsed().as_nanos() as u64;
-        let mut cpu_t = self
-            .world
-            .resource_mut::<byroredux_core::ecs::CpuFrameTimings>();
-        cpu_t.atw_pre_ms = atw_pre_ns as f32 * NS_TO_MS;
-        cpu_t.atw_scheduler_ms = atw_scheduler_ns as f32 * NS_TO_MS;
-        cpu_t.atw_post_ms = atw_post_ns as f32 * NS_TO_MS;
-    }
-}
 
 /// Phase 4b — build the per-frame [`PanelSnapshot`] the egui
 /// overlay reads. Always populates `metrics`; the entity list is
@@ -2067,81 +672,5 @@ fn apply_debug_ui_outputs(
         for line in response_lines {
             ui.push_console_line(line);
         }
-    }
-}
-
-/// Regression for #2522 / PERF-D6-NEW-01. `draw_frame` can return `Err`
-/// from at least four sites (fence wait, command-buffer reset/begin, FSR
-/// parameter build) that all execute before `record_skinned_blas_refit` —
-/// the call that flips `skin_dispatch_ran` true. The `skin_dispatch_ran`
-/// rollback check must therefore run unconditionally on the `Result`, not
-/// nested inside the `Ok` match arm, or an early `Err` silently loses the
-/// #1791/#1796 first-sight-upload/pose-hash rollback. A live `App`/
-/// `VulkanContext` test is impractical here for the same reason
-/// `skin_dispatch_ran_ordering_tests` in `draw.rs` gives (70+ Vulkan-loader
-/// fields, no safe defaults) — a static source assertion pins the ordering
-/// instead, mirroring that sibling test's technique on the caller side.
-#[cfg(test)]
-mod skin_dispatch_ran_rollback_scope_tests {
-    #[test]
-    fn rollback_check_runs_before_the_ok_err_match_not_inside_the_ok_arm() {
-        let src = include_str!("main.rs");
-
-        let draw_call_pos = src
-            .find("let draw_result = ctx.draw_frame(FrameInputs {")
-            .expect("render_one_frame must call draw_frame and capture its Result (#2522)");
-        let rollback_check_pos = src
-            .find("if !ctx.skin_dispatch_ran {")
-            .expect("render_one_frame must check skin_dispatch_ran for rollback (#1791/#1796)");
-        let match_pos = src
-            .find("match draw_result {")
-            .expect("render_one_frame must match on the captured draw_result (#2522)");
-        let ok_arm_pos = src
-            .find("Ok(needs_recreate) => {")
-            .expect("draw_result match must have an Ok(needs_recreate) arm");
-
-        assert!(
-            draw_call_pos < rollback_check_pos,
-            "the skin_dispatch_ran rollback check must come AFTER the \
-             draw_frame call, so it observes this frame's outcome. (#2522)"
-        );
-        assert!(
-            rollback_check_pos < match_pos,
-            "the skin_dispatch_ran rollback check must come BEFORE the \
-             match on draw_result — i.e. outside and above both arms — or \
-             the Err(e) arm would skip it entirely, silently losing the \
-             #1791/#1796 rollback on any of draw_frame's early-Err paths. \
-             (#2522)"
-        );
-        assert!(
-            match_pos < ok_arm_pos,
-            "sanity: the Ok(needs_recreate) arm must be part of the \
-             draw_result match this test is reasoning about."
-        );
-    }
-}
-
-/// A deterministic bench must remember the scene/CLI camera before the
-/// startup scheduler runs `camera_follow_system`. The per-frame reapply is too
-/// late to recover the authored origin if this ordering regresses.
-#[cfg(test)]
-mod bench_camera_startup_order_tests {
-    #[test]
-    fn authored_bench_origin_is_seeded_before_startup_scheduler() {
-        let src = include_str!("main.rs");
-        let setup = src
-            .find("self.setup_scene();")
-            .expect("renderer startup must set up the scene");
-        let startup = &src[setup..];
-        let seed = startup
-            .find("self.seed_bench_camera_origin();")
-            .expect("startup must preserve the authored bench camera");
-        let scheduler = startup
-            .find("self.scheduler.run(&self.world, 0.0);")
-            .expect("startup must prime the scheduler");
-        assert!(
-            seed < scheduler,
-            "bench origin must be captured before character camera sync can overwrite it"
-        );
     }
 }
